Add route guard service
