--- src/app/core/route-guard/route-guard.service.ts
@@ -1,15 +1,0 @@
-import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, CanActivateChild, Router, RouterStateSnapshot } from '@angular/router';
-
-@Injectable()
-export class RouteGuardService implements CanActivate, CanActivateChild {
-    constructor(private router: Router) {}
-
-    async canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<boolean> {
-        return true;
-    }
-
-    async canActivateChild(childRoute: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<boolean> {
-        return true;
-    }
-}
